Changed ino_t size to 32 bits

+++ newkern/include/crt/sys/types.h
@@ -35,7 +35,7 @@ typedef int pid_t;
 typedef unsigned short uid_t;
 typedef unsigned short gid_t;
 typedef unsigned short dev_t;
-typedef unsigned short ino_t;
+typedef unsigned long ino_t;
 typedef unsigned long mode_t;
 typedef unsigned long umode_t;
 typedef unsigned short nlink_t;
